+++ models/sectionTypes/Text.php
@@ -78,11 +78,10 @@ abstract class Text extends ISectionType
         $str .= HTMLTools::prepareHTMLForCkeditor($data);
         $str .= '</div>';
 
-        if (HTMLTools::cleanSimpleHtml($originalHtml) !== HTMLTools::cleanSimpleHtml($data)) {
-            $str .= '<div class="modifiedActions"><button class="btn-link resetText" type="button">';
-            $str .= \Yii::t('amend', 'revert_changes');
-            $str .= '</button></div>';
-        }
+        $changed = HTMLTools::cleanSimpleHtml($originalHtml) !== HTMLTools::cleanSimpleHtml($data);
+        $str .= '<div class="modifiedActions' . ($changed ? '' : ' hidden') . '"><button class="btn-link resetText" type="button">';
+        $str .= \Yii::t('amend', 'revert_changes');
+        $str .= '</button></div>';
 
         $str .= '</div>';